switch to pyproject macros (rhbz#2378537)
dblatex uses a peculiar setup.py where build_scripts uses options to the
install command in order to generate the dblatex script. This happens
during `setup.py install` but never with the new macros. We solve this
by patching the proper paths directly into setup.py. An upstreamable fix
would need to rewrite setup.py's logic or switch to pyproject.tompl
completely.

Along the way, we remove shebangs from some non-scripts. They used to be
rewritten (and would not any more) but should not be there in the first
place.